don't try to update if chirper is hidden'

+++ Birdcage/BirdcageUserMod.cs
@@ -106,6 +106,11 @@
                     return;
                 }
 
+                if (ModConfig.Instance.GetSetting<bool>(SettingKeys.HideChirper))
+                {
+                    return;
+                }
+
                 if (ModConfig.Instance.GetSetting<bool>(SettingKeys.FilterMessages))
                 {
                     filterService.RemovePendingMessages(NotificationSound);
